requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-login==0.6.2
werkzeug==2.3.7
python-dateutil==2.8.2
flask-pymongo==2.3.0 